Quizz_IA/requirements.txt
Flask~=3.1.0
Flask-JWT-Extended~=4.7.1
Werkzeug~=3.1.3
googletrans~=4.0.0rc1
transformers~=4.47.1
gTTS~=2.5.4
langdetect~=1.0.9
deepface~=0.0.93
Flask-Cors~=5.0.0
python-dotenv~=1.0.1
chardet~=3.0.4
PyPDF2~=3.0.1
python-docx~=1.1.2
pyttsx3~=2.98